website/requirements.txt
django==5.1.11
djangorestframework==3.15.2
psycopg2-binary==2.9.9
django-cors-headers==4.4.0
django-extensions==3.2.3
numpy==1.26.4
ortools==9.10.4067